style(web): page loads now
