+++ libs/hooks/src/index.ts
@@ -1,1 +1,1 @@
-export { usePrefix } from './lib/use-prefix/use-prefix';
+export * from './lib/use-supabase-client/use-supabase-client';
